yaffs mount on mtd works

+++ bl2/incl/types.h
@@ -5,7 +5,8 @@ typedef unsigned long int size_t;
 typedef long int ssize_t;
 
 typedef unsigned long int off_t;
-typedef unsigned long long int loff_t;
+typedef unsigned long int loff_t;
+//typedef unsigned long long int loff_t;
 
 typedef unsigned int dev_t;
